Add dark mode and routing with sortable tables
