+++ frontend/package.json
@@ -13,7 +13,8 @@
     "clsx": "^2.1.0",
     "tailwind-merge": "^2.2.1",
     "class-variance-authority": "^0.7.0",
-    "lucide-react": "^0.379.0"
+    "lucide-react": "^0.379.0",
+    "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^4.0.0",
